requirements_simple.txt
# Simple Gemini Document Processor
# Minimal dependencies for direct image processing

google-generativeai>=0.8.0
pillow>=10.0.0
python-dotenv>=1.0.0

# Optional: Web interface
flask>=2.3.0